EntireDocument->FileDocument save()->write() Uses Files.class to read/write deleted StringBuilder instance
